+++ lib/shape/ellipse.dart
@@ -1,11 +1,7 @@
-import 'dart:math' as Math;
-
 import 'package:voxel_generator/shape/base.dart';
+import 'package:voxel_generator/shape/shape_utils.dart';
 
 class Ellipse extends Shape3d {
-  double get a => halfWidth;
-  double get b => halfDepth;
-
   Ellipse({
     int width,
     int height,
@@ -16,8 +12,5 @@ class Ellipse extends Shape3d {
   );
 
   @override
-  bool contains(num x, num y, num z) {
-    return (Math.pow(x, 2) / Math.pow(a, 2))
-      + (Math.pow(y, 2) / Math.pow(b, 2)) <= 1;
-  }
+  bool contains(num x, num y, num z) => ShapeUtils.pointInEllipse(x, y, halfWidth, halfDepth);
 }